Refactor: splitting Hwnd
