+++ tests/flutter_engine/flutter_finder_test.dart
@@ -1,0 +1,5 @@
+// search relative path currdir, workdir
+
+// search absolute path currdir, workdir
+
+// TODO: flutter finder tests
